fixed type error in calculate_progression

+++ src/pathogen/progression_categories/pc_asymptomatic.jl
@@ -36,10 +36,10 @@ end
 function calculate_progression(individual::Individual, tick::Int16, dp::Asymptomatic, rng::Xoshiro)
         
     # Calculate the time to infectiousness
-    infectiousness_onset = tick + Int16(1) + rand_val(dp.exposure_to_infectiousness_onset, rng)
+    infectiousness_onset = Int16(tick + 1 + rand_val(dp.exposure_to_infectiousness_onset, rng))
 
     # Calculate the time to recovery
-    recovery = infectiousness_onset + rand_val(dp.infectiousness_onset_to_recovery, rng)
+    recovery = Int16(infectiousness_onset + rand_val(dp.infectiousness_onset_to_recovery, rng))
 
     return DiseaseProgression(
         exposure = tick,
