+++ src/pathogen/progression_categories/pc_severe.jl
@@ -45,19 +45,19 @@ end
 function calculate_progression(individual::Individual, tick::Int16, dp::Severe, rng::Xoshiro)
 
     # Calculate the time to infectiousness
-    infectiousness_onset = tick + Int16(1) + rand_val(dp.exposure_to_infectiousness_onset, rng)
+    infectiousness_onset = Int16(tick + 1 + rand_val(dp.exposure_to_infectiousness_onset, rng))
 
     # Calculate the time to symptom onset
-    symptom_onset = infectiousness_onset + rand_val(dp.infectiousness_onset_to_symptom_onset, rng)
+    symptom_onset = Int16(infectiousness_onset + rand_val(dp.infectiousness_onset_to_symptom_onset, rng))
 
     # Calculate the time to severeness onset
-    severeness_onset = symptom_onset + rand_val(dp.symptom_onset_to_severeness_onset, rng)
+    severeness_onset = Int16(symptom_onset + rand_val(dp.symptom_onset_to_severeness_onset, rng))
 
     # Calculate the time to severeness offset
-    severeness_offset = severeness_onset + rand_val(dp.severeness_onset_to_severeness_offset, rng)
+    severeness_offset = Int16(severeness_onset + rand_val(dp.severeness_onset_to_severeness_offset, rng))
 
     # Calculate the time to recovery
-    recovery = severeness_offset + rand_val(dp.severeness_offset_to_recovery, rng)
+    recovery = Int16(severeness_offset + rand_val(dp.severeness_offset_to_recovery, rng))
 
     return DiseaseProgression(
         exposure = tick,
